requirements.txt
numpy
opencv-contrib-python
opencv-python
openpyxl
pandas
pillow
pyttsx3
flask>=2.0.0
flask-cors>=3.0.0
qrcode[pil]>=7.0.0
reportlab>=3.6.0
psycopg2-binary
sqlalchemy>=1.4.0